Corrected to working state
Added the wait prior to removing the `local eventData`, as this was having a scheduler error on interpreted output, it couldn't deal with its own variable being removed inside itself.
Amended the `TriggerClientCallback` as this was preventing 0 aka "the server" as the source. Permitting ">= 0" allows the server to pass requests without error.